pyproject.toml
[project]
name = "video-analysis-phase3"
version = "0.1.0"
description = "Phase3 local-first video analysis stack (API + Worker + MCP)"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.116.0,<1",
  "google-genai>=1.33.0,<2",
  "httpx>=0.28.1,<1",
  "markdown>=3.8.2,<4",
  "mcp>=1.13.1,<2",
  "psycopg[binary]>=3.2.10,<4",
  "pydantic>=2.11.7,<3",
  "sqlalchemy>=2.0.43,<3",
  "temporalio>=1.17.0,<2",
  "tzdata>=2025.2",
  "uvicorn>=0.35.0,<1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1,<9",
  "pytest-cov>=6.0.0,<7",
  "pytest-rerunfailures>=15.1,<16",
  "pytest-xdist>=3.8.0,<4",
  "mutmut>=3.3.1,<4",
]
e2e = [
  "playwright>=1.55.0,<2",
]

[tool.pytest.ini_options]
pythonpath = [
  ".",
  "apps/worker",
]
reruns = 2
reruns_delay = 1

[tool.ruff]
target-version = "py311"
line-length = 100
fix = true

[tool.ruff.lint]
select = ["ALL"]
ignore = [
  "ANN",
  "D",
  "COM812",
  "ISC001",
  "EM",
  "TRY",
  "TC",
  "TID",
  "PLR",
  "S",
  "FAST",
  "B008",
  "T20",
  "ASYNC",
  "FBT",
  "BLE",
  "FURB",
  "E501",
  "A",
  "C901",
  "DTZ",
  "PGH",
  "RUF",
  "PERF",
  "INP",
  "PTH",
  "N",
  "PT",
  "ARG",
  "UP038",
  "PLC0415",
  "EXE001",
  "PLW2901",
  "RET504",
  "SLF001",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.per-file-ignores]
"**/tests/**/*.py" = ["S101", "S105"]
"**/test_*.py" = ["S101", "S105"]

[tool.mutmut]
debug = false
paths_to_mutate = [
  "apps/worker/worker/pipeline/steps/llm_step_gates.py",
]
also_copy = [
  "apps/worker/worker",
  "apps/worker/tests",
]
max_stack_depth = 32
pytest_add_cli_args_test_selection = [
  "apps/worker/tests/test_llm_step_gates.py",
]
